+++ src/experimental/include/exaDEM/experimental/polyhedron/nbh_gpu/nbh_manager.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <exaDEM/experimental/polyhedron/nbh_gpu/nbh_gpu_driver.hpp>
+#include <exaDEM/traversal.hpp>
 
 namespace exaDEM {
 struct NBHManager {
@@ -9,11 +10,8 @@ struct NBHManager {
   CellDriverStorage info_cell_driver;
 };
 
-inline void classify_interaction_grid(
-    Classifier& classifier,
-    Traversal& traversal,
-    NBHManager& nbh_manager,
-    GridCellParticleInteraction& ges) {
+inline void classify_interaction_grid(Classifier& classifier, Traversal& traversal, NBHManager& nbh_manager,
+                                      GridCellParticleInteraction& ges) {
   InteractionWrapperStorage wrappers(classifier);
   InteractionWrapperAccessor wrapper_accessor = wrappers.accessor();
   auto [cell_ptr, cell_size] = traversal.info();
@@ -22,8 +20,6 @@ inline void classify_interaction_grid(
   constexpr bool do_active_interaction_only = true;
 
   transfer_classifier_grid<do_ghost_only, do_active_interaction_only>(
-      cell_ptr, nbh_manager.info_cell, nbh_manager.info_pair_cell,
-       nbh_manager.info_cell_driver, wrapper_accessor, ges);
-
+      cell_ptr, nbh_manager.info_cell, nbh_manager.info_pair_cell, nbh_manager.info_cell_driver, wrapper_accessor, ges);
 }
 }  // namespace exaDEM
